Fix ingestion sampling (#3578)

+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/sampling/AiSampler.java
@@ -76,6 +76,10 @@ public class AiSampler implements Sampler {
               : parentlessDependencySamplingPercentage.get();
     }
 
+    if (sp == SamplingPercentage.USE_INGESTION_SAMPLING) {
+      return SamplingResult.recordAndSample();
+    }
+
     if (sp == 0) {
       return SamplingResult.drop();
     }
